+++ AudioKit/Common/Nodes/Generators/Oscillators/Oscillator Bank/AKOscillatorBankAudioUnit.h
@@ -14,9 +14,9 @@
 
 - (void)setupWaveform:(int)size;
 - (void)setWaveformValue:(float)value atIndex:(UInt32)index;
-- (void)startNote:(uint8_t)note velocity:(uint8_t)velocity;
-- (void)startNote:(uint8_t)note velocity:(uint8_t)velocity frequency:(float)frequency;
-- (void)stopNote:(uint8_t)note;
+- (void)startNote:(uint16_t)note velocity:(uint8_t)velocity;
+- (void)startNote:(uint16_t)note velocity:(uint8_t)velocity frequency:(float)frequency;
+- (void)stopNote:(uint16_t)note;
 - (void)reset;
 
 @end
